+++ android/src/main/kotlin/cl/ceisufro/native_video_view/NativeVideoViewFactory.kt
@@ -1,0 +1,18 @@
+package cl.ceisufro.native_video_view
+
+import android.content.Context
+import io.flutter.plugin.common.PluginRegistry
+import io.flutter.plugin.common.StandardMessageCodec
+import io.flutter.plugin.platform.PlatformView
+import io.flutter.plugin.platform.PlatformViewFactory
+import java.util.concurrent.atomic.AtomicInteger
+
+class NativeVideoViewFactory(private val activityState: AtomicInteger,
+                             private val registrar: PluginRegistry.Registrar)
+    : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
+
+    override fun create(context: Context?, id: Int, args: Any?): PlatformView {
+        val builder = NativeVideoViewBuilder()
+        return builder.build(id, activityState, registrar)
+    }
+}