+++ example/lib/main.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:native_video_view/native_video_view.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(MaterialApp(home: MyApp()));
 
 class MyApp extends StatefulWidget {
   @override
@@ -12,44 +9,37 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await NativeVideoView.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
-        ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
-        ),
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Plugin example app'),
+      ),
+      body: _buildVideoPlayerWidget(),
+    );
+  }
+
+  Widget _buildVideoPlayerWidget() {
+    return Container(
+      alignment: Alignment.center,
+      child: NativeVideoView(
+        keepAspectRatio: true,
+        showMediaController: true,
+        onCreated: (controller) {
+          controller.setVideoSource(
+            'assets/example.mp4',
+            sourceType: VideoSourceType.asset,
+          );
+        },
+        onPrepared: (controller, info) {
+          controller.play();
+        },
+        onError: (controller, what, extra, message) {
+          print('Player Error ($what | $extra | $message)');
+        },
+        onCompletion: (controller) {
+          print('Video completed');
+        },
       ),
     );
   }