+++ lib/src/media_control.dart
@@ -1,0 +1,10 @@
+part of native_video_view;
+
+/// Enum to specify the types of controls in the media controller.
+enum _MediaControl {
+  pause,
+  play,
+  stop,
+  fwd,
+  rwd,
+}