+++ test/native_video_view_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:native_video_view/native_video_view.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('native_video_view');
@@ -13,9 +12,5 @@ void main() {
 
   tearDown(() {
     channel.setMockMethodCallHandler(null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await NativeVideoView.platformVersion, '42');
   });
 }